old projekt/remote desktop/requirements.txt
pyautogui==0.9.54
mss==9.0.1
keyboard==0.13.5
Pillow==10.1.0
numpy==1.26.2
